- manage.py script added - an option --new added (for creating new file for a problem).
I am just bored to write filename for every problem. From now I can just copy problem's url, paste it to the terminal and hit enter and new file will be created and opened in my editor (and there will be a couple of common imports and a class Solution with a method).